+++ src/pybispectra/general/__init__.py
@@ -1,5 +1,5 @@
 """Tools for handling general bispectrum and threenorm computations."""
 
-__version__ = "1.2.0dev"
+__version__ = "1.2.0"
 
 from .general import Bispectrum, Threenorm